fix: isolate structured sidecar manifest types
