+++ src/core/extension/mod.rs
@@ -15,6 +15,7 @@ mod maintenance;
 mod manifest;
 mod manifest_action_config;
 mod manifest_config;
+mod manifest_sidecar;
 mod refactor_protocol;
 mod registry;
 mod repair;
